embedding_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sentence-transformers==3.0.1
torch==2.1.2
transformers==4.36.2
numpy==1.24.3
huggingface_hub==0.23.5